全体災害とセルイベントをEvent配下に移動
