+++ app/app/Entity/Event/Disaster/AppearanceMonster.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Entity\Disaster;
+namespace App\Entity\Event\Disaster;
 
 use App\Entity\Cell\Cell;
 use App\Entity\Cell\CellTypeConst;
