+++ app/app/Entity/Event/ForeignIsland/ForeignIslandEvent.php
@@ -1,27 +1,27 @@
 <?php
 
-namespace App\Entity\Plan\ForeignIsland\Plan;
+namespace App\Entity\Event\ForeignIsland;
 
-use App\Entity\Plan\Plan;
+use App\Entity\Cell\Cell;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 use App\Models\Island;
 use App\Models\Turn;
 
-abstract class TargetedToForeignIslandPlan
+abstract class ForeignIslandEvent
 {
     protected int $fromIsland;
     protected int $toIsland;
-    protected Plan $plan;
+    protected Cell $cell;
 
-    public function __construct(int $fromIsland, int $toIsland, Plan $plan)
+    public function __construct(int $fromIsland, int $toIsland, Cell $cell)
     {
         $this->fromIsland = $fromIsland;
         $this->toIsland = $toIsland;
-        $this->plan = $plan;
+        $this->cell = $cell;
     }
 
-    public abstract function execute(Island $fromIsland, Island $toIsland, Terrain $fromTerrain, Terrain $toTerrain, Status $fromStatus, Status $toStatus, Turn $turn): ExecutePlanToForeignIslandResult;
+    public abstract function execute(Island $fromIsland, ?Island $toIsland, Terrain $fromTerrain, ?Terrain $toTerrain, Status $fromStatus, ?Status $toStatus, Turn $turn): ForeignIslandEventResult;
 
     /**
      * @return int
@@ -40,10 +40,10 @@ abstract class TargetedToForeignIslandPlan
     }
 
     /**
-     * @return Plan
+     * @return Cell
      */
-    public function getPlan(): Plan
+    public function getCell(): Cell
     {
-        return $this->plan;
+        return $this->cell;
     }
 }
