+++ app/app/Entity/Plan/ForeignIsland/ExecutePlanToForeignIslandResult.php
@@ -1,21 +1,21 @@
 <?php
 
-namespace App\Entity\Plan\ForeignIsland\Event;
+namespace App\Entity\Plan\ForeignIsland;
 
 use App\Entity\Log\Logs;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 
-class ForeignIslandOccurEventResult
+class ExecutePlanToForeignIslandResult
 {
     private Terrain $fromTerrain;
-    private ?Terrain $toTerrain;
+    private Terrain $toTerrain;
     private Status $fromStatus;
-    private ?Status $toStatus;
+    private Status $toStatus;
     private Logs $fromLogs;
     private Logs $toLogs;
 
-    public function __construct(Terrain $fromTerrain, ?Terrain $toTerrain, Status $fromStatus, ?Status $toStatus, Logs $fromLogs, ?Logs $toLogs)
+    public function __construct(Terrain $fromTerrain, Terrain $toTerrain, Status $fromStatus, Status $toStatus, Logs $fromLogs, Logs $toLogs)
     {
         $this->fromTerrain = $fromTerrain;
         $this->toTerrain = $toTerrain;
@@ -34,9 +34,9 @@ class ForeignIslandOccurEventResult
     }
 
     /**
-     * @return Status|null
+     * @return Status
      */
-    public function getToStatus(): ?Status
+    public function getToStatus(): Status
     {
         return $this->toStatus;
     }
@@ -50,9 +50,9 @@ class ForeignIslandOccurEventResult
     }
 
     /**
-     * @return Terrain|null
+     * @return Terrain
      */
-    public function getToTerrain(): ?Terrain
+    public function getToTerrain(): Terrain
     {
         return $this->toTerrain;
     }
@@ -66,9 +66,9 @@ class ForeignIslandOccurEventResult
     }
 
     /**
-     * @return Logs|null
+     * @return Logs
      */
-    public function getToLogs(): ?Logs
+    public function getToLogs(): Logs
     {
         return $this->toLogs;
     }
