+++ app/app/Entity/Plan/ForeignIsland/TargetedToForeignIslandPlan.php
@@ -1,27 +1,27 @@
 <?php
 
-namespace App\Entity\Plan\ForeignIsland\Event;
+namespace App\Entity\Plan\ForeignIsland;
 
-use App\Entity\Cell\Cell;
+use App\Entity\Plan\Plan;
 use App\Entity\Status\Status;
 use App\Entity\Terrain\Terrain;
 use App\Models\Island;
 use App\Models\Turn;
 
-abstract class ForeignIslandOccurEvent
+abstract class TargetedToForeignIslandPlan
 {
     protected int $fromIsland;
     protected int $toIsland;
-    protected Cell $cell;
+    protected Plan $plan;
 
-    public function __construct(int $fromIsland, int $toIsland, Cell $cell)
+    public function __construct(int $fromIsland, int $toIsland, Plan $plan)
     {
         $this->fromIsland = $fromIsland;
         $this->toIsland = $toIsland;
-        $this->cell = $cell;
+        $this->plan = $plan;
     }
 
-    public abstract function execute(Island $fromIsland, ?Island $toIsland, Terrain $fromTerrain, ?Terrain $toTerrain, Status $fromStatus, ?Status $toStatus, Turn $turn): ForeignIslandOccurEventResult;
+    public abstract function execute(Island $fromIsland, Island $toIsland, Terrain $fromTerrain, Terrain $toTerrain, Status $fromStatus, Status $toStatus, Turn $turn): ExecutePlanToForeignIslandResult;
 
     /**
      * @return int
@@ -40,10 +40,10 @@ abstract class ForeignIslandOccurEvent
     }
 
     /**
-     * @return Cell
+     * @return Plan
      */
-    public function getCell(): Cell
+    public function getPlan(): Plan
     {
-        return $this->cell;
+        return $this->plan;
     }
 }
