+++ app/app/Entity/Terrain/Terrain.php
@@ -17,9 +17,9 @@ use App\Entity\Cell\Sea;
 use App\Entity\Cell\Shallow;
 use App\Entity\Cell\Volcano;
 use App\Entity\Cell\Wasteland;
-use App\Entity\Disaster\DisasterConst;
-use App\Entity\Disaster\DisasterResult;
-use App\Entity\Disaster\IDisaster;
+use App\Entity\Event\Disaster\DisasterConst;
+use App\Entity\Event\Disaster\DisasterResult;
+use App\Entity\Event\Disaster\IDisaster;
 use App\Entity\Log\Logs;
 use App\Entity\Status\Status;
 use App\Entity\Util\Normal;
